WIP modal, also refactor all global modals
