--- src/app/(authenticated)/(globalModals)/layout.tsx
@@ -1,18 +1,0 @@
-import type { FC, PropsWithChildren } from 'react';
-
-import AddFolderModal from '@/app/_modals/AddFolderModal';
-import AddTalkModal from '@/app/_modals/AddTalkModal';
-import ConfirmationModal from '@/app/_modals/ConfirmationModal';
-import RootFolderErrorModal from '@/app/_modals/RootFolderErrorModal';
-
-const Layout: FC<PropsWithChildren> = ({ children }) => (
-    <>
-        {children}
-        <AddTalkModal />
-        <AddFolderModal />
-        <ConfirmationModal />
-        <RootFolderErrorModal />
-    </>
-);
-
-export default Layout;
