+++ src/app/(authenticated)/talks/[slug]/_components/TalkAttribute/index.tsx
