+++ src/app/(authenticated)/talks/scan/page.tsx
